# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Medical Scheduling System API
  description: |
    A comprehensive medical scheduling system for hospital nurses to automate recurring test and injection schedules.
    Built with Next.js 15, TypeScript, and Supabase.

    **Recent Updates** (2025-11-10, PR 50):
    - Improved error handling: 404 responses for missing resources (previously 500)
    - Department API now uses .maybeSingle() for proper RESTful error codes
  version: 1.0.1
  contact:
    name: API Support
    email: support@medical-scheduler.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.medical-scheduler.com
    description: Production server
security:
  - BearerAuth: []
...
paths:
...
  /admin/delete-user:
    post:
      tags:
        - Admin
      summary: Delete user (2-Layer Architecture)
      description: |
        Permanently delete a user account using simplified 2-layer architecture.

        **Architecture (Phase 8 - Nov 2025)**:
        - Layer 1: API Route → Single Supabase Auth API call
        - Layer 2: Database → Automatic triggers and FK CASCADE

        **Automatic Processing**:
        1. BEFORE DELETE Triggers:
           - `check_last_admin()`: Prevents deletion of last admin
           - `anonymize_user_audit_logs()`: Preserves audit trail
        2. FK CASCADE: Auto-cleanup (notifications, schedule references)
        3. AFTER DELETE Triggers: All active (data integrity)

        **Requirements**:
        - Admin role required
        - Cannot delete own account
        - Cannot delete last admin (database-level protection)

        **Data Preservation**:
        - Audit logs anonymized (not deleted)
        - Schedule history preserved with NULL references
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
              properties:
                userId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User deleted successfully
                  userId:
                    type: string
                    format: uuid
        '400':
          description: Bad request - Cannot delete own account or last admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - Admin access required or invalid origin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Target user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
  responses:
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            code: AUTH_REQUIRED
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error
            code: INTERNAL_ERROR
...